This is what I have for bank right now
I have added the changes that I made to other classes as well. The bank works mostly aside from a few bugs.
The bugs are:
1. The first car tends to pick up the most passengers
2. The average longest wait time is dependent on the number of cars in the bank
as well as the run time, this isn't that big of a bug but when there are ten cars for ten floors and 
it takes a while it can be.